Integrate more async policy implementation
