+++ server/service/service_osquery.go
@@ -778,7 +778,7 @@ func (svc *Service) SubmitDistributedQueryResults(
 		// TODO(mna): should async host processing overrule deferred save host? It
 		// is strictly "more async". We could pass the deferred save host setting to
 		// task.RecordLabelQueryExecutions and let it dispatch appropriately (to async
-		// or datastore).
+		// or datastore) - I've applied this approach below.
 
 		//if ac.ServerSettings.DeferredSaveHost {
 		//	if err := svc.ds.RecordLabelQueryExecutions(ctx, &host, labelResults, svc.clock.Now(), true); err != nil {
@@ -807,14 +807,13 @@ func (svc *Service) SubmitDistributedQueryResults(
 		}
 		// TODO(mna): is it an issue for the webhook if the policy executions are
 		// temporarily recorded in redis? Obviously it wouldn't see the new flipped
-		// policies on the next run, but would it typically run before the redis async
-		// data is collected and persisted in mysql? Should FlippingPoliciesForHost
-		// take pending redis data into consideration? Should we impose restrictions
-		// between async collection interval and policy update interval?
+		// policies on the next run, but would it typically run before the redis
+		// async data is collected and persisted in mysql? Should
+		// FlippingPoliciesForHost take pending redis data into consideration, or
+		// maybe we should impose restrictions between async collection interval
+		// and policy update interval?
 
-		host.PolicyUpdatedAt = svc.clock.Now()
-		err = svc.ds.RecordPolicyQueryExecutions(ctx, &host, policyResults, svc.clock.Now(), ac.ServerSettings.DeferredSaveHost)
-		if err != nil {
+		if err := svc.task.RecordPolicyQueryExecutions(ctx, &host, policyResults, svc.clock.Now(), ac.ServerSettings.DeferredSaveHost); err != nil {
 			logging.WithErr(ctx, err)
 		}
 	}
